Final/requirements.txt
fastapi
uvicorn
streamlit
pydantic
python-dotenv
openai
langchain
langgraph
faiss-cpu
pandas
